requirements.txt
python-keystoneclient==3.11.0
python-cinderclient==2.2.0
python-novaclient<=3.3.3
